requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
tensorflow==2.20.0
Pillow==12.0.0
numpy==2.3.5
scipy==1.16.3
kagglehub==0.3.13
werkzeug==3.1.4
scikit-learn==1.6.0
matplotlib==3.10.0